pyproject.toml
[project]
name = "llm-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "crewai>=0.159.0",
    "jsonpatch>=1.33",
    "jsonpointer>=3.0.0",
    "langchain>=0.3.27",
    "langchain-google-genai>=2.1.9",
    "langchain-mistralai>=0.2.11",
    "langchain-ollama>=0.3.7",
    "langchain-openai>=0.3.31",
    "llama-index>=0.13.3",
    "llama-index-llms-google-genai>=0.3.0",
    "llama-index-llms-mistralai>=0.7.0",
    "llama-index-llms-ollama>=0.7.1",
    "openai>=1.101.0",
    "openai-agents[litellm]>=0.2.9",
    "pydantic>=2.11.7",
    "pydantic-ai>=0.7.5",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "pytest>=8.4.1",
]
